# Avoid select/cat/unbind in AffineAR inverse